// src/app/interfaces/user.ts
export interface User {
  id?: string;
  email?: string;
  password?: string;
  nome?: string;
  nome_mae?: string;
  data_nascimento?: string;
  cpf?: string;
  rg?: string;
  cep?: string;
  uf?: string;
  logradouro?: string;
  complemento?: string;
  numero?: string;
  telefone?: string;
  celular?: string;
  createdAt?: Date;
  level?: string;
}